+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
+- Added: Included the ALB arns (api and push) as outputs
+- Updated: Fixed "time_zone" parameter value - used incorrect value previously
+- Added: Added parameter "security_code_removal_mins" to control number of minutes before a one-time upload code is removed from the database by the cleanup lambda
+- Added: Added parameter "upload_max_keys" to control maximum number of keys accepted per upload request
 - Added: Added option to configure SNS SMS spent quota
 - Added: Added option to configure SNS SMS delivery logs